+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog of reinout-media
 0.1 (unreleased)
 ----------------
 
-- Added Photo datamodel.
+- Added Tree datamodel: wraps a directory with photo files.
+
+- Added Photo datamodel: wraps a photo file.
 
 - Initial project structure created with nensskel 1.30.